api/api.py
from io import BytesIO

from utils import http_get
from PIL import Image


brawlers_url = "https://api.brawlapi.com/v1/brawlers"
brawlers_data = http_get(brawlers_url).json()['list']

for brawler_obj in brawlers_data:
    icon_url = brawler_obj['imageUrl2']
    response = http_get(icon_url)
    image = Image.open(BytesIO(response.content))
    image.save(f"./assets/brawler_icons2/{str(brawler_obj['name']).lower()}.png")
