requirements.txt
asyncio~=3.4.3
customtkinter~=5.2.2
opencv-python~=4.11.0.86
numpy~=2.3.0
onnxruntime-directml
pyautogui~=0.9.54
pygetwindow~=0.0.9
httpx~=0.28.1
toml~=0.10.2
torch
pillow~=11.2.1
shapely~=2.1.1
ultralytics~=8.3.233
bettercam~=1.0.0
packaging~=25.0
pywin32
pure-python-adb
google-play-scraper~=1.2.7
easyocr~=1.7.2
scrcpy-client@git+https://github.com/leng-yue/py-scrcpy-client.git@v0.5.0
adbutils>=2.0.0
